requirements.txt
a2wsgi==1.10.0
amazon-braket-default-simulator==1.20.1
amazon-braket-schemas==1.19.1.post0
amazon-braket-sdk==1.64.1
antlr4-python3-runtime==4.9.2
anyio==4.2.0
appdirs==1.4.4
Arpeggio==2.0.2
asgiref==3.7.2
attrs==23.1.0
autograd==1.6.2
autoray==0.6.7
backoff==2.2.1
blinker==1.7.0
boltons==23.1.1
boto3==1.34.2
boto3-stubs==1.34.2
botocore==1.34.2
botocore-stubs==1.34.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cirq-core==1.3.0
cirq-google==1.3.0
click==8.1.7
cloudpickle==2.2.1
clyngor==0.4.3
connexion==3.0.5
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.14
dictances==1.5.6
dill==0.3.7
duet==0.2.9
Flask==3.0.0
fonttools==4.46.0
future==0.18.3
google-api-core==2.10.2
google-auth==2.25.2
googleapis-common-protos==1.62.0
graphviz==0.20.1
grpcio==1.60.0
grpcio-status==1.48.2
h11==0.14.0
httpcore==0.16.3
httptools==0.6.1
httpx==0.23.3
hypothesis==6.92.1
ibm-cloud-sdk-core==3.18.2
ibm-platform-services==0.48.1
idna==3.6
importlib-metadata==7.0.1
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
lark==0.11.3
lark-parser==0.12.0
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mpmath==1.3.0
msgpack==1.0.7
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
ntlm-auth==1.5.0
numpy==1.23.5
openpulse==0.4.2
openqasm3==0.4.0
opt-einsum==3.3.0
oqpy==0.2.1
packaging==23.2
pandas==2.1.4
pbr==6.0.0
PennyLane==0.33.1
PennyLane-Lightning==0.33.1
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
ply==3.11
projectq==0.8.0
proto-plus==1.23.0
protobuf==3.20.3
psutil==5.9.7
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
pyparsing==3.1.1
pyPEG2==2.15.2
pyqir==0.9.0
pyquil==3.5.4
pyspnego==0.10.2
pytest==7.4.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-rapidjson==1.14
pytket==1.22.0
pytket-braket==0.32.0
pytket-cirq==0.32.0
pytket-pennylane==0.13.0
pytket-projectq==0.31.0
pytket-pyquil==0.31.0
pytket-qir==0.6.0
pytket-qiskit==0.46.0
pytket-qsharp==0.37.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
qcs-api-client==0.21.6
qiskit==0.45.1
qiskit-aer==0.13.1
qiskit-algorithms==0.2.1
qiskit-ibm-provider==0.7.3
qiskit-ibm-runtime==0.15.1
qiskit-ibmq-provider==0.20.2
qiskit-ionq==0.4.6
qiskit-terra==0.45.1
qsharp==0.28.302812
qsharp-chemistry==0.28.302812
qsharp-core==0.28.302812
qwasm==1.0.1
referencing==0.32.0
requests==2.31.0
requests-ntlm==1.1.0
retry==0.9.2
retrying==1.3.4
rfc3339==6.2
rfc3986==1.5.0
rpcq==3.11.0
rpds-py==0.15.2
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
rustworkx==0.13.2
s3transfer==0.9.0
scipy==1.11.4
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.32.0.post1
statistical-distance==0.9.1
stevedore==5.1.0
support-developer==1.0.5
swagger_ui_bundle==1.1.0
symengine==0.9.2
sympy==1.12
tenacity==8.2.3
toml==0.10.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
types-awscrt==0.20.0
types-Deprecated==1.2.9.3
types-pkg-resources==0.1.3
types-python-dateutil==2.8.19.14
types-requests==2.31.0.10
types-retry==0.9.9.4
types-s3transfer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.25.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
